refactor -- remove the unnecessary `src/compile/stack.ts` and its empty test file.
