+++ src/compile/mark/mark.ts
@@ -1,18 +1,20 @@
-import {UnitModel} from '../unit';
-import {OrderChannelDef} from '../../fielddef';
-
-import {X, Y, COLOR, TEXT, SHAPE, PATH, ORDER, OPACITY, DETAIL, LABEL} from '../../channel';
+import {X, Y, COLOR, TEXT, SHAPE, PATH, ORDER, OPACITY, DETAIL, LABEL, STACK_GROUP_CHANNELS} from '../../channel';
+import {has} from '../../encoding';
+import {OrderChannelDef, FieldDef, field} from '../../fielddef';
 import {AREA, LINE, TEXT as TEXTMARK} from '../../mark';
-import {stackTransforms} from '../stack';
-import {contains, extend} from '../../util';
+import {ScaleType} from '../../scale';
+import {contains, extend, isArray} from '../../util';
+import {VgStackTransform} from '../../vega.schema';
+
 import {area} from './area';
 import {bar} from './bar';
+import {sortField} from '../common';
 import {line} from './line';
 import {point, circle, square} from './point';
+import {rule} from './rule';
 import {text} from './text';
 import {tick} from './tick';
-import {rule} from './rule';
-import {sortField} from '../common';
+import {UnitModel} from '../unit';
 
 const markCompiler = {
   area: area,
@@ -209,4 +211,80 @@ function detailFields(model: UnitModel): string[] {
     }
     return details;
   }, []);
+}
+
+
+function stackTransforms(model: UnitModel, impute: boolean): any[] {
+  const stackByFields = getStackByFields(model);
+  if (impute) {
+    return [imputeTransform(model, stackByFields), stackTransform(model, stackByFields)];
+  }
+  return [stackTransform(model, stackByFields)];
+}
+
+
+/** Compile stack-by field names from (from 'color' and 'detail') */
+function getStackByFields(model: UnitModel) {
+  const encoding = model.encoding();
+
+  return STACK_GROUP_CHANNELS.reduce(function(fields, channel) {
+    const channelEncoding = encoding[channel];
+    if (has(encoding, channel)) {
+      if (isArray(channelEncoding)) {
+        channelEncoding.forEach(function(fieldDef) {
+          fields.push(field(fieldDef));
+        });
+      } else {
+        const fieldDef: FieldDef = channelEncoding;
+        const scale = model.scale(channel);
+        fields.push(field(fieldDef, {
+          binSuffix: scale && scale.type === ScaleType.ORDINAL ? '_range' : '_start'
+        }));
+      }
+    }
+    return fields;
+  }, []);
+}
+
+// impute data for stacked area
+function imputeTransform(model: UnitModel, stackFields: string[]) {
+  const stack = model.stack();
+  return {
+    type: 'impute',
+    field: model.field(stack.fieldChannel),
+    groupby: stackFields,
+    orderby: [model.field(stack.groupbyChannel, {binSuffix: '_mid'})],
+    method: 'value',
+    value: 0
+  };
+}
+
+function stackTransform(model: UnitModel, stackFields: string[]) {
+  const stack = model.stack();
+  const encoding = model.encoding();
+  const sortby = model.has(ORDER) ?
+    (isArray(encoding[ORDER]) ? encoding[ORDER] : [encoding[ORDER]]).map(sortField) :
+    // default = descending by stackFields
+    stackFields.map(function(field) {
+     return '-' + field;
+    });
+
+  const valName = model.field(stack.fieldChannel);
+
+  // add stack transform to mark
+  let transform: VgStackTransform = {
+    type: 'stack',
+    groupby: [model.field(stack.groupbyChannel, {binSuffix: '_mid'})],
+    field: model.field(stack.fieldChannel),
+    sortby: sortby,
+    output: {
+      start: valName + '_start',
+      end: valName + '_end'
+    }
+  };
+
+  if (stack.offset) {
+    transform.offset = stack.offset;
+  }
+  return transform;
 }
