+++ src/vega.schema.ts
@@ -82,3 +82,12 @@ export type VgMarkGroup = any;
 export type VgAxis = any;
 export type VgLegend = any;
 export type VgTransform = any;
+
+export interface VgStackTransform {
+  type: string;
+  offset?: any;
+  groupby: any;
+  field: any;
+  sortby: any;
+  output: any;
+}
