--- test/compile/stack.test.ts
@@ -1,3 +1,0 @@
-/* tslint:disable:quotemark */
-
-// TODO: write properly
