+++ tsconfig.json
@@ -66,7 +66,6 @@
     "src/compile/mark/tick.ts",
     "src/compile/model.ts",
     "src/compile/scale.ts",
-    "src/compile/stack.ts",
     "src/compile/time.ts",
     "src/compile/unit.ts",
     "src/config.ts",
@@ -103,7 +102,6 @@
     "test/compile/mark/text.test.ts",
     "test/compile/mark/tick.test.ts",
     "test/compile/scale.test.ts",
-    "test/compile/stack.test.ts",
     "test/compile/unit.test.ts",
     "test/fielddef.test.ts",
     "test/schema.test.ts",
